factor: move IR dump to separate module
